Add files via upload
I have made some modifications to the solidworks macro for exporting STL files to Cura. Currently every file that is sent to Cura has the same name (model_for_cura). This isn't very helpful when you want to print of an SD card. My fix fetches the part name and exports the file to cura using this part name.